Added a few comments to silence warnings and changed temp to a float64.

+++ pkg/messaging/telemetry/telemetry.go
@@ -2,15 +2,18 @@ package telemetry
 
 import "github.com/iot-for-tillgenglighet/ingress-trafikverket/pkg/messaging"
 
+// Temperature is a telemetry type IoTHubMessage
 type Temperature struct {
 	messaging.IoTHubMessage
-	Temp float32 `json:"temp"`
+	Temp float64 `json:"temp"`
 }
 
+// ContentType returns the ContentType for a Temperature telemetry message
 func (msg *Temperature) ContentType() string {
 	return "application/json"
 }
 
+// TopicName returns the name of the topic that a Temperature telemetry message should be posted to
 func (msg *Temperature) TopicName() string {
 	return "telemetry.temperature"
 }
